Cropping disabled by default
A lot of times, the cropping function cuts relevants part of the comic. I would prefer this function to be disabled or at least have the possibility to switch on/off with a click on the GUI. If that isn't possible, can i have a compiled version with this change? Or maybe the instruction to compile it myself. Thanks. Sorry for the English.